+++ src/main/java/ma/vi/esql/lookup/JoinLabel.java
@@ -51,6 +51,11 @@ import static ma.vi.esql.translation.Translatable.Target.ESQL;
  * </p>
  *
  * <p>
+ * When the code to get the label for is `null`, joinlabel loads all possible
+ * labels. This can be used to populate a drop-down for selection.
+ * </p>
+ *
+ * <p>
  * joinlabel can have the following optional named arguments to control the
  * value displayed:
  * <ul>
@@ -62,8 +67,13 @@ import static ma.vi.esql.translation.Translatable.Target.ESQL;
  *                           last linked table to the first, if true, or otherwise,
  *                           from the first to the last. Default is true.</li>
  * <li><b>matching:</b> A criteria to restrict the code-label pairs to load from
- *                      the lookup. Applies only when the code searched is null,
- *                      meaning that the whole lookup is to be loaded.</li>
+ *                      the lookup. Applies only when the code searched is null.</li>
+ * <li><b>labels_offset:</b> An offset from the start of loaded data from where
+ *                           to start returning labels. Applies only when the code
+ *                           searched is null; can be used to lazily load labels.</li>
+ * <li><b>labels_limit:</b> The number labels to return. Applies only when the
+ *                          code searched is null; can be used to lazily load
+ *                          labels in pages.</li>
  * </ul>
  * @author Vikash Madhow (vikash.madhow@gmail.com)
  */
@@ -104,6 +114,8 @@ public class JoinLabel extends Function implements TypedMacro {
     Expression<?, ?> labelSeparator = new StringLiteral(ctx, " / ");    // the separator to use between labels from different table (joins).
     boolean          lastToFirst    = true;                             // show labels last to first (or first to last if false).
     String           matching       = null;                             // Criteria to restrict code-label pairs to load for whole lookup.
+    Expression<?, ?> offset         = null;                             // Offset labels loading by this number.
+    Expression<?, ?> limit          = null;                             // Limit labels to load to this number.
 
     Iterator<Expression<?, ?>> i = arguments.iterator();
     while (i.hasNext()) {
@@ -114,6 +126,8 @@ public class JoinLabel extends Function implements TypedMacro {
           case "last_to_first"   -> lastToFirst    = getBooleanParam(namedArg, "last_to_first", path);
           case "label_separator" -> labelSeparator = namedArg.arg();
           case "matching"        -> matching       = getStringParam(namedArg, "matching", path);
+          case "labels_offset"   -> offset         = namedArg.arg();
+          case "labels_limit"    -> limit          = namedArg.arg();
           default                -> throw new TranslationException("Invalid named argument in joinlabel: " + namedArg.name());
         }
       } else {
@@ -207,6 +221,12 @@ public class JoinLabel extends Function implements TypedMacro {
       if (matching != null) {
         Expression<?, String> where = ColumnRef.qualify(builder.parser.parseExpression(matching), firstFromAlias);
         builder.where(where);
+      }
+      if (offset != null) {
+        builder.offset((Expression<?, String>)offset);
+      }
+      if (limit != null) {
+        builder.limit((Expression<?, String>)limit);
       }
       return builder.column (firstTargetId, "code")
                     .column (value, "label")